Refactor Poincaré algorithms to align with center manifold structure
- Updated import paths to reflect the new organization of center manifold components, including changes from `_shbackend` and `_cmbackend` to their respective new locations.
- Removed deprecated files and classes related to the old Poincaré structure, streamlining the codebase.
- Enhanced the clarity and maintainability of the code by cleaning up import statements and adjusting method signatures.
- Ensured consistency across the codebase by updating references to the new center manifold configuration and backend classes.
