+++ src/hiten/algorithms/poincare/cm/backend.py
@@ -11,8 +11,8 @@ from hiten.algorithms.integrators.rk import (RK4_A, RK4_B, RK4_C, RK6_A, RK6_B,
                                              RK6_C, RK8_A, RK8_B, RK8_C)
 from hiten.algorithms.integrators.symplectic import (N_SYMPLECTIC_DOF,
                                                      _integrate_symplectic)
-from hiten.algorithms.poincare.backends.base import _ReturnMapBackend
-from hiten.algorithms.poincare.config import _get_section_config
+from hiten.algorithms.poincare.core.backend import _ReturnMapBackend
+from hiten.algorithms.poincare.cm.config import _get_section_config
 from hiten.algorithms.poincare.events import _SurfaceEvent
 from hiten.algorithms.polynomial.operations import _polynomial_evaluate
 from hiten.algorithms.utils.config import FASTMATH
