+++ src/hiten/algorithms/poincare/cm/config.py
@@ -13,25 +13,14 @@ from typing import Literal, Optional, Tuple
 
 import numpy as np
 
+from hiten.algorithms.poincare.core.config import (_ReturnMapConfig,
+                                                   _SectionConfig)
 from hiten.utils.log_config import logger
 
 
-class _Section:
-    def __init__(self, pts, st, labels):
-        self.points = pts
-        self.states = st
-        self.labels = labels
-
-
 @dataclass
-class _CenterManifoldMapConfig:
-    dt: float = 1e-2
-    method: Literal["scipy", "rk", "symplectic", "adaptive"] = "rk"
-    order: int = 4
-    c_omega_heuristic: float = 20.0
+class _CenterManifoldMapConfig(_ReturnMapConfig):
 
-    n_seeds: int = 20
-    n_iter: int = 40
     seed_strategy: Literal[
         "single",
         "axis_aligned",
@@ -39,10 +28,9 @@ class _CenterManifoldMapConfig:
         "radial",
         "random",
     ] = "axis_aligned"
+
     seed_axis: Optional[Literal["q2", "p2", "q3", "p3"]] = None
     section_coord: Literal["q2", "p2", "q3", "p3"] = "q3"
-
-    compute_on_init: bool = False
 
     def __post_init__(self):
         if self.seed_strategy == "single" and self.seed_axis is None:
@@ -51,7 +39,7 @@ class _CenterManifoldMapConfig:
             logger.warning("seed_axis is ignored when seed_strategy is not 'single'")
 
 
-class _CenterManifoldSectionConfig:
+class _CenterManifoldSectionConfig(_SectionConfig):
 
     _TABLE: dict[str, dict[str, object]] = {
         "q3": dict(
