+++ src/hiten/algorithms/poincare/core/backend.py
