+++ src/hiten/algorithms/poincare/core/base.py
@@ -1,0 +1,8 @@
+
+class _Section:
+    def __init__(self, points, states, labels):
+        self.points = points
+        self.states = states
+        self.labels = labels
+
+
