+++ src/hiten/system/orbits/halo.py
@@ -25,7 +25,7 @@ from typing import TYPE_CHECKING, Literal, Optional, Sequence
 import numpy as np
 from numpy.typing import NDArray
 
-from hiten.algorithms.poincare.backends._shbackend import _y_plane_crossing
+from hiten.algorithms.poincare.sh.backend import _y_plane_crossing
 from hiten.system.libration.base import LibrationPoint
 from hiten.system.libration.collinear import (CollinearPoint, L1Point, L2Point,
                                               L3Point)
